Mv attestation stuff to attestation module
