+++ crates/threshold-signature-server/src/lib.rs
@@ -122,6 +122,7 @@
 //!     [sled](https://docs.rs/sled)
 #![doc(html_logo_url = "https://entropy.xyz/assets/logo_02.png")]
 pub use entropy_client::chain_api;
+pub(crate) mod attestation;
 pub(crate) mod health;
 pub mod helpers;
 pub(crate) mod node_info;
@@ -149,13 +150,14 @@ pub use crate::helpers::{
     validator::{get_signer, get_signer_and_x25519_secret},
 };
 use crate::{
+    attestation::api::attest,
     health::api::healthz,
     launch::Configuration,
     node_info::api::{hashes, version as get_version},
     r#unsafe::api::{delete, put, remove_keys, unsafe_get},
     signing_client::{api::*, ListenerState},
     user::api::*,
-    validator::api::{attest, new_reshare},
+    validator::api::new_reshare,
 };
 
 #[derive(Clone)]
