+++ crates/threshold-signature-server/src/validator/api.rs
@@ -360,54 +360,3 @@ pub async fn prune_old_holders(
         validators_info.clone()
     })
 }
-
-/// HTTP POST endpoint to initiate a TDX attestation.
-/// Not yet implemented.
-#[cfg(not(any(test, feature = "unsafe")))]
-pub async fn attest(
-    State(_app_state): State<AppState>,
-    _input: Bytes,
-) -> Result<StatusCode, ValidatorErr> {
-    // Non-mock attestation (the real thing) will go here
-    Err(ValidatorErr::NotImplemented)
-}
-
-/// HTTP POST endpoint to initiate a mock TDX attestation for testing on non-TDX hardware.
-/// The body of the request should be a 32 byte random nonce used to show 'freshness' of the
-/// quote.
-/// The response body contains a mock TDX v4 quote serialized as described in the
-/// [Index TDX DCAP Quoting Library API](https://download.01.org/intel-sgx/latest/dcap-latest/linux/docs/Intel_TDX_DCAP_Quoting_Library_API.pdf).
-#[cfg(any(test, feature = "unsafe"))]
-pub async fn attest(
-    State(app_state): State<AppState>,
-    input: Bytes,
-) -> Result<(StatusCode, Bytes), ValidatorErr> {
-    // TODO (#982) confirm with the chain that an attestation should be happenning
-    let nonce = input.as_ref().try_into()?;
-
-    let rpc = get_rpc(&app_state.configuration.endpoint).await?;
-
-    let block_number = rpc
-        .chain_get_header(None)
-        .await?
-        .ok_or_else(|| ValidatorErr::OptionUnwrapError("Failed to get block number".to_string()))?
-        .number;
-
-    // In the real thing this is the hardware key used in the quoting enclave
-    let signing_key = tdx_quote::SigningKey::random(&mut OsRng);
-
-    let (signer, x25519_secret) = get_signer_and_x25519_secret(&app_state.kv_store).await?;
-    let public_key = x25519_dalek::PublicKey::from(&x25519_secret);
-
-    let input_data = entropy_shared::QuoteInputData::new(
-        signer.signer().public().into(),
-        *public_key.as_bytes(),
-        nonce,
-        block_number,
-    );
-
-    let quote = tdx_quote::Quote::mock(signing_key.clone(), input_data.0);
-    // Here we would submit an attest extrinsic to the chain - but for now we just include it in the
-    // response
-    Ok((StatusCode::OK, Bytes::from(quote.as_bytes().to_vec())))
-}
